@@ -7,6 +7,7 @@
     "build": "ng build",
     "watch": "ng build --watch --configuration development",
     "test": "ng test --code-coverage",
+    "test-ci": "ng test --no-watch --no-progress --browsers=ChromeHeadless",
     "http-server": "npx http-server -p 8080 -c-1 dist/spiewnigator-angular/browser",
     "build-deploy": "ng build --base-href=/spiewnigator-dev/",
     "lint": "ng lint",
@@ -14,41 +15,41 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^18.2.8",
-    "@angular/cdk": "^18.2.8",
-    "@angular/common": "^18.2.8",
-    "@angular/compiler": "^18.2.8",
-    "@angular/core": "^18.2.8",
-    "@angular/forms": "^18.2.8",
-    "@angular/material": "^18.2.8",
-    "@angular/platform-browser": "^18.2.8",
-    "@angular/platform-browser-dynamic": "^18.2.8",
-    "@angular/router": "^18.2.8",
-    "@angular/service-worker": "^18.2.8",
+    "@angular/animations": "^20.1.3",
+    "@angular/cdk": "^20.1.3",
+    "@angular/common": "^20.1.3",
+    "@angular/compiler": "^20.1.3",
+    "@angular/core": "^20.1.3",
+    "@angular/forms": "^20.1.3",
+    "@angular/material": "^20.1.3",
+    "@angular/platform-browser": "^20.1.3",
+    "@angular/platform-browser-dynamic": "^20.1.3",
+    "@angular/router": "^20.1.3",
+    "@angular/service-worker": "^20.1.3",
     "rxjs": "~6.6.0",
     "tslib": "^2.6.3",
-    "zone.js": "^0.14.10"
+    "zone.js": "^0.15.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^18.2.8",
-    "@angular-eslint/builder": "18.3.1",
-    "@angular-eslint/eslint-plugin": "18.3.1",
-    "@angular-eslint/eslint-plugin-template": "18.3.1",
-    "@angular-eslint/schematics": "18.3.1",
-    "@angular-eslint/template-parser": "18.3.1",
-    "@angular/cli": "^18.2.8",
-    "@angular/compiler-cli": "^18.2.8",
+    "@angular-eslint/builder": "20.1.1",
+    "@angular-eslint/eslint-plugin": "20.1.1",
+    "@angular-eslint/eslint-plugin-template": "20.1.1",
+    "@angular-eslint/schematics": "20.1.1",
+    "@angular-eslint/template-parser": "20.1.1",
+    "@angular/build": "^20.1.3",
+    "@angular/cli": "^20.1.3",
+    "@angular/compiler-cli": "^20.1.3",
     "@types/jasmine": "~3.8.0",
     "@types/node": "^12.20.23",
-    "@typescript-eslint/eslint-plugin": "^7.2.0",
-    "@typescript-eslint/parser": "^7.2.0",
+    "@typescript-eslint/eslint-plugin": "^8.38.0",
+    "@typescript-eslint/parser": "^8.38.0",
     "eslint": "^8.57.0",
     "jasmine-core": "~3.8.0",
-    "karma": "~6.3.0",
+    "karma": "^6.4.0",
     "karma-chrome-launcher": "~3.1.0",
     "karma-coverage": "^2.2.1",
     "karma-jasmine": "~4.0.0",
     "karma-jasmine-html-reporter": "~1.7.0",
-    "typescript": "~5.4.5"
+    "typescript": "~5.8.3"
   }
 }
